+++ myokit/_aux.py
@@ -434,6 +434,45 @@ def python_writer():
     return _pywriter_
 
 
+def _prepare_bindings(model, labels):
+    """
+    Takes a mapping from labels to strings and returns a mapping from variable
+    objects to strings, for all labels that are defined in the model.
+
+    Raises an error if the label refers to a variable that is not a literal.
+
+    The argument ``mapping`` should take the form::
+
+        labels = {
+            'binding_label_1' : internal_name_1,
+            'binding_label_2' : internal_name_2,
+            ...
+            }
+
+    The returned dictionary will have the form::
+
+        variables = {
+            variable_x : internal_name_1,
+            variable_y : internal_name_2,
+            ...
+            }
+
+    Unsupported bindings (i.e. bindings not appearing in ``labels``) will
+    be ignored.
+    """
+    unused = []
+    variables = {}
+    for label, var in self._bindings.items():
+        try:
+            variables[var] = labels[label]
+        except KeyError:
+            unused.append(var)
+            continue
+        var.set_rhs(0)
+    for var in unused:
+        var.set_binding(None)
+    return variables
+
 def run(model, protocol, script, stdout=None, stderr=None, progress=None):
     """
     Runs a python ``script`` using the given ``model`` and ``protocol``.